requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python
scikit-learn
matplotlib
tqdm
pyyaml
streamlit
